+++ app/__init__.py
@@ -1,15 +1,15 @@
 from flask import Flask, jsonify
 from app import search, get_detail_common
 import config
-from app.views.api import api
-from app.views.upload import upload
+from app.view.api import view
+from app.upload.api import upload
 from flask_cors import CORS
 
 
 def create_app(config_file=config.Config):
     app = Flask(__name__)
     app.config.from_object(config_file)
-    app.register_blueprint(api)
+    app.register_blueprint(view)
     app.register_blueprint(upload)
     CORS(app)
